Update CatalogPage to take in columns
